requirements.txt
accelerate==1.4.0
bitsandbytes==0.45.3
faiss-gpu==1.7.2
huggingface-hub==0.29.1
ipykernel==6.23.3
peft==0.14.0
PyPDF2==3.0.1
pandas==2.0.2
numpy==1.23.5
sentence-transformers==3.4.1
streamlit==1.43.1
torch==2.6.0
tqdm==4.67.1
transformers==4.46.1
trl==0.15.2